# back-end/docs/swagger.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/auth/traveler-login": {
      "post": {
        "description": "Authenticate a traveler using email and password from in-memory data.",
        "operationId": "AuthController_travelerLogin",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TravelerLoginDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Login successful",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/LoginResponseDto"
                }
              }
            }
          },
          "400": {
            "description": "Validation error — email or password missing/invalid"
          },
          "401": {
            "description": "Invalid email or password"
          }
        },
        "summary": "Traveler login",
        "tags": [
          "auth"
        ]
      }
    },
    "/auth/staff-login": {
      "post": {
        "description": "Authenticate admin, superuser, guide, or support staff using email and password.",
        "operationId": "AuthController_staffLogin",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/StaffLoginDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Login successful",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/LoginResponseDto"
                }
              }
            }
          },
          "400": {
            "description": "Validation error — email or password missing/invalid"
          },
          "401": {
            "description": "Invalid email or password"
          }
        },
        "summary": "Staff login",
        "tags": [
          "auth"
        ]
      }
    },
    "/auth/agency-login": {
      "post": {
        "description": "Authenticate an agency using email and password. Fails if agency is not approved.",
        "operationId": "AuthController_agencyLogin",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AgencyLoginDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Login successful",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/LoginResponseDto"
                }
              }
            }
          },
          "400": {
            "description": "Validation error — email or password missing/invalid"
          },
          "401": {
            "description": "Invalid email, password, or agency not approved"
          }
        },
        "summary": "Agency login",
        "tags": [
          "auth"
        ]
      }
    },
...
    "/agencies/register": {
      "post": {
        "operationId": "AgenciesController_create",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateAgencyDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Agency registered successfully (pending status)"
          },
          "400": {
            "description": "Validation error / Email in use",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseDto"
                }
              }
            }
          }
        },
        "summary": "Register a new agency (public)",
        "tags": [
          "agencies"
        ]
      }
    },
    "/agencies": {
      "post": {
        "operationId": "AgenciesController_createDirect",
        "parameters": [
          {
            "name": "x-user-role",
            "in": "header",
            "description": "User role for RBAC. Allowed values: traveler, guide, support, admin, superuser, agency",
            "required": false,
            "schema": {
              "type": "string",
              "enum": [
                "traveler",
                "guide",
                "support",
                "admin",
                "superuser",
                "agency"
              ]
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateDirectAgencyDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Agency created"
          },
          "400": {
            "description": "Validation error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Create a new agency directly (Admin/Superuser only)",
        "tags": [
          "agencies"
        ]
      },
      "get": {
        "operationId": "AgenciesController_findAll",
        "parameters": [
          {
            "name": "x-user-role",
            "in": "header",
            "description": "User role for RBAC. Allowed values: traveler, guide, support, admin, superuser, agency",
            "required": false,
            "schema": {
              "type": "string",
              "enum": [
                "traveler",
                "guide",
                "support",
                "admin",
                "superuser",
                "agency"
              ]
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Return all agencies"
          },
          "403": {
            "description": "Forbidden",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Get all agencies (Admin/Superuser only)",
        "tags": [
          "agencies"
        ]
      }
    },
    "/agencies/{id}": {
      "get": {
        "operationId": "AgenciesController_findOne",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "number"
            }
          },
          {
            "name": "x-user-role",
            "in": "header",
            "description": "User role for RBAC. Allowed values: traveler, guide, support, admin, superuser, agency",
            "required": false,
            "schema": {
              "type": "string",
              "enum": [
                "traveler",
                "guide",
                "support",
                "admin",
                "superuser",
                "agency"
              ]
            }
          },
          {
            "name": "x-user-id",
            "in": "header",
            "description": "User ID for ownership verification",
            "required": false,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Return the agency"
          },
          "403": {
            "description": "Forbidden (not owner)",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseDto"
                }
              }
            }
          },
          "404": {
            "description": "Not Found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Get a specific agency by ID",
        "tags": [
          "agencies"
        ]
      },
      "patch": {
        "operationId": "AgenciesController_update",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "number"
            }
          },
          {
            "name": "x-user-role",
            "in": "header",
            "description": "User role for RBAC. Allowed values: traveler, guide, support, admin, superuser, agency",
            "required": false,
            "schema": {
              "type": "string",
              "enum": [
                "traveler",
                "guide",
                "support",
                "admin",
                "superuser",
                "agency"
              ]
            }
          },
          {
            "name": "x-user-id",
            "in": "header",
            "description": "User ID for ownership verification",
            "required": false,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateAgencyDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Agency updated"
          },
          "403": {
            "description": "Forbidden",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseDto"
                }
              }
            }
          },
          "404": {
            "description": "Not Found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Update agency profile (Agency only ownership checked, or Admin/Superuser)",
        "tags": [
          "agencies"
        ]
      },
      "delete": {
        "operationId": "AgenciesController_remove",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "number"
            }
          },
          {
            "name": "x-user-role",
            "in": "header",
            "description": "User role for RBAC. Allowed values: traveler, guide, support, admin, superuser, agency",
            "required": false,
            "schema": {
              "type": "string",
              "enum": [
                "traveler",
                "guide",
                "support",
                "admin",
                "superuser",
                "agency"
              ]
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Agency deleted"
          },
          "404": {
            "description": "Not Found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Delete an agency (Admin/Superuser only)",
        "tags": [
          "agencies"
        ]
      }
    },
    "/agencies/{id}/status": {
      "patch": {
        "operationId": "AgenciesController_updateStatus",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "number"
            }
          },
          {
            "name": "x-user-role",
            "in": "header",
            "description": "User role for RBAC. Allowed values: traveler, guide, support, admin, superuser, agency",
            "required": false,
            "schema": {
              "type": "string",
              "enum": [
                "traveler",
                "guide",
                "support",
                "admin",
                "superuser",
                "agency"
              ]
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateAgencyStatusDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Status updated"
          },
          "403": {
            "description": "Forbidden",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseDto"
                }
              }
            }
          },
          "404": {
            "description": "Not Found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Update agency status & commission (Admin/Superuser only)",
        "tags": [
          "agencies"
        ]
      }
    },
...
  },
  "info": {
    "title": "GlobeSync API",
    "description": "GlobeSync Travel Platform Backend API",
    "version": "1.0.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "schemas": {
      "TravelerLoginDto": {
        "type": "object",
        "properties": {
          "email": {
            "type": "string",
            "description": "Traveler email address",
            "example": "Arjun@gmail.com"
          },
          "password": {
            "type": "string",
            "description": "Traveler password",
            "example": "traveler123"
          }
        },
        "required": [
          "email",
          "password"
        ]
      },
      "LoginResponseDto": {
        "type": "object",
        "properties": {
          "message": {
            "type": "string",
            "example": "Login successful"
          },
          "role": {
            "type": "string",
            "example": "traveler",
            "enum": [
              "traveler",
              "admin",
              "superuser",
              "guide",
              "support"
            ]
          },
          "user": {
            "type": "object",
            "description": "User object (shape depends on role)",
            "example": {
              "travelerId": 1,
              "name": "Arjun Mehta",
              "email": "Arjun@gmail.com",
              "phone": "9876543210",
              "gender": "Male",
              "age": 28
            }
          }
        },
        "required": [
          "message",
          "role",
          "user"
        ]
      },
      "StaffLoginDto": {
        "type": "object",
        "properties": {
          "email": {
            "type": "string",
            "description": "Staff email address",
            "example": "admin@ticp.com"
          },
          "password": {
            "type": "string",
            "description": "Staff password",
            "example": "admin123"
          }
        },
        "required": [
          "email",
          "password"
        ]
      },
      "AgencyLoginDto": {
        "type": "object",
        "properties": {
          "email": {
            "type": "string",
            "example": "contact@elitetravel.com"
          },
          "password": {
            "type": "string",
            "example": "password"
          }
        },
        "required": [
          "email",
          "password"
        ]
      },
...
      "CreateAgencyDto": {
        "type": "object",
        "properties": {
          "businessName": {
            "type": "string",
            "example": "Elite Travel Partners"
          },
          "contactEmail": {
            "type": "string",
            "example": "contact@elitetravel.com"
          },
          "contactPhone": {
            "type": "string",
            "example": "9876543210"
          },
          "password": {
            "type": "string",
            "example": "securepassword123"
          }
        },
        "required": [
          "businessName",
          "contactEmail",
          "contactPhone",
          "password"
        ]
      },
      "ErrorResponseDto": {
        "type": "object",
        "properties": {
          "statusCode": {
            "type": "number",
            "example": 404
          },
          "error": {
            "type": "string",
            "example": "Not Found"
          },
          "message": {
            "type": "object",
            "example": "Resource with the given ID was not found"
          },
          "timestamp": {
            "type": "string",
            "example": "2026-05-05T12:00:00.000Z"
          },
          "path": {
            "type": "string",
            "example": "/travelers/abc-123"
          }
        },
        "required": [
          "statusCode",
          "error",
          "message",
          "timestamp",
          "path"
        ]
      },
      "CreateDirectAgencyDto": {
        "type": "object",
        "properties": {
          "businessName": {
            "type": "string",
            "example": "Elite Travel Partners"
          },
          "contactEmail": {
            "type": "string",
            "example": "contact@elitetravel.com"
          },
          "contactPhone": {
            "type": "string",
            "example": "9876543210"
          },
          "password": {
            "type": "string",
            "example": "securepassword123"
          },
          "status": {
            "type": "string",
            "example": "approved",
            "enum": [
              "pending",
              "approved",
              "rejected"
            ]
          },
          "commissionRate": {
            "type": "number",
            "example": 10
          }
        },
        "required": [
          "businessName",
          "contactEmail",
          "contactPhone",
          "password"
        ]
      },
      "UpdateAgencyStatusDto": {
        "type": "object",
        "properties": {
          "status": {
            "type": "string",
            "example": "approved",
            "enum": [
              "pending",
              "approved",
              "rejected"
            ]
          },
          "commissionRate": {
            "type": "number",
            "example": 15,
            "description": "Commission rate percentage, typically required when approving"
          }
        },
        "required": [
          "status"
        ]
      },
      "UpdateAgencyDto": {
        "type": "object",
        "properties": {
          "businessName": {
            "type": "string",
            "example": "Elite Travel Partners"
          },
          "contactEmail": {
            "type": "string",
            "example": "contact@elitetravel.com"
          },
          "contactPhone": {
            "type": "string",
            "example": "9876543210"
          },
          "password": {
            "type": "string",
            "example": "securepassword123"
          }
        }
      },
...
    }
  }
}